added default executable path. allows to call OUR functions (echo, pwd, etc) via execve and not just simple fork. why? no fucking idea, just because i can.
